dockercomposes/requirements.txt
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
Flask-Compress==1.8.0
fsspec==0.7.4
influxdb==5.3.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.5.2
numpy==1.22.0
pandas==1.3.5
pandocfilters==1.4.2
plotly
sqlalchemy==1.4.29
psycopg2==2.9.3
pytz==2021.3
python-dotenv==0.19.2
    